// swiyu-ecosystem-portal-ui/src/app/api/generated/model/signatory.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Signatory {
  firstName: string;
  lastName: string;
  phone: string;
  email: string;
}